llm_client/requirements.txt
requests
pymongo
kafka-python